refactor/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
iniconfig==2.0.0
itsdangerous==2.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mistune==3.0.2
mysqlclient==2.2.6
packaging==24.2
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
PyMySQL==1.1.1

